Remove assigment behaviour
I do this carefully and via a dedicated commit. The insert behaviour now
supported via the subset helper seems to be in too strong overlap with
assignment for us to support both. I will explicitly discuss with
original author simon heybrock.

Also note that no tests break as a result of this removal. Does not seem
to be heavily used.