Dataset binary (non in-place) operators
Added plus, minus, times and divide operators.

When non-trivial to implement, rvalue parameterised operations are also
provided.

Note that the times operator taking const lvalue LHS and rvalue RHS is
not provided, difference in order of operands caused this to give
slightly different results to other times operators for Dataset.