Move Docker guide to Read the Docs
Moves the using dataset in Docker guide from a readme to the Read the
Docs documentation.

Skips duplicate documentation already in the Installation RTD page and
details on building the container (building is planned to be automatic
and users wishing to build manually will probably already know
everything that section mentioned).