Include demos and exmaples in image
Includes a copy of the demo notebooks and Python examples from the
current repository state at image build time.

The ability to update these is not included in the image, this adds
complexity when simply updating the image would provide the desired
outcome.
Given that the last two layers are installing Scipp and obtaining the
demos there should be minimal time/network traffic involved in users
using the image in this way.