Use unordered_map instead of map.
The reason is to avoid client code (unit tests) depending on a specific
item order, which is not guaranteed (but if we return sorted items code
is starting to rely on that and will be harder to change later).

As a side-effect, we move away from std::string_view since
std::unordered_map does not support transparent comparators (may be more
convenient also if we eventually switch to an optmized map
implementation). This was a bit messy anyway.