Added convert_TableWorkspace_to_dataset and included it in load wrapper function.
All data is loaded as Dim.Row and unitless since no information on
units is available in the Mantid TableWorkspace.
The TableWorkspace does however have some limited information about
which column is data and which is the error on that data. If the
included data provides a clear mapping between data and error, this
is used and the generated scipp variables will have the appropriate
variance. This data can however be ambigious, in which case a
runtime error is raised. A keyword argument called error_connection
is availble to manually specify the name of the assosiated error
colum for each data column. If the keyword argument is given, no
automatic recoignition of the included plottype data is performed.

A blacklist of types is available, but it seems scipp can handle
all available datatypes so it is currently empty.

Further documentation is available as a docstring for the function.