Simple implementation of diffraction converter from time of flight to dspacing using a calibration file. Only the linear terms are currently used, but it can easily be expanded two include the quadratic term in the future.
The convert_with_calibration function is in a new namespace under
neutron called diffraction as it is specific to this technique.

The function takes two datasets, one with the data to be used and
another with the calibration data for each detector. The Dims for
the calibration data and data needs to match, with the exception
of the actual Tof dimension.

It is the intention to raise errors when the required fields are
not available in the given calibration data, but the necessary
exceptions did not work as expected. The calls are commented out.

The function was tested with a modified version of the dream.py
code found in scipp/ess and it performed as expected. This code
will be uploaded on Jira under das-123.