Consider multiple workspaces in Mantid loader
When loading certain types of data (e.g. event data from NeXus files)
multiple workspaces may be produced (one for data and another for
monitors).

This adds checking of the return type and handles additional workspaces.

Monitors are not actually used at this point, but loading no longer
fails if you ask for them to be loaded.

Fixes #677
Re #615