Add config file mechanics
Adds ability to load config files and retrieve values.

Config file is read form sensible location for the current platform.
If it does not exist it is created with the defaults.

On loading, a configuration based on the defaults is updated with values
form the user config file to form the runtime config.

Note that additions to the default config will not be automatically
populated in the user configuration file. This is closer to what
existing tools would do.

Re #758