Remove Doxygen from docs build
Doxygen is not used and does not produce anything particularly useful.
This commit removes it for the documentation build.

Removes:
  - Doxygen dependency
  - Doxygen config
  - Doxygen build step
  - References to Doxygen docs in main docs

Doxygen comments remain and should be kept as up to date as before.
If the need arises it can be re-added at a later date.

Also tidies up the variable construction documentation page.