Fix variable tests asserting now unrelated base class
Fixes variable tests which throw a DimensionMismatchError. This is
aliased as MimatchError (which is a derived from Error). However the tests
caught a DimensionError (which was typedef'd from Error). Now that the
DimensionError is a derived type, instead of a type of errot the
inheritance chain no longer matches DimensionMismatchError as a type of
DimensionError.

We fix this up by specifying in the tests that it's a
DimensionMismatchError were expecting, as that is what's thrown anyway