Move plotting implementation to impl module
Moves the plotting implementation to an impl module. This clearly
seperates a public and private API acting as a boundary for what people
should be calling. For example it avoids people trying to force their
data through plot_1d when they should be instead passing a 1d dataset
(or using collapse) with plot