Add tiled plotting API to scipp.plot
Adds a tiled plotting API to scipp plot file. This packs into a list of
requested plots which are plotted as tiles.

For 1d/2d slices the figure creation was moved out of the slicer widget
to the caller. This way the caller can create a widget per subplot.

3d plots just spawn multiple widgets within the same cell since
ipyvolume handles its own window management