Fixes Fit WS Index not handling float values
The compat layer can accept a float point value for spectrum numbers,
however Mantid will throw a really unhelpful exception if it tries to
use a float instead of an int. Convert in the shim layer so users get a
useful message if it's not int compatible, or it's handled silently for
them.