Replace bitmask comparisons with logical comparitors
Replaces bitmask comparisons with logical comparitors. MSVC warns that
we are taking T and T& and comparing T with the pointer rather than the
expected T with T.

Additionally, this resolves the fact that a&b != a&&b if upper bits
differ, for example if the compiler does not set them to 0 as an
optimisation. From a user perspective a binary representation of 01 and
11 will both be True, but 01 != 11 using the & operator