Use tag dispatch to get ElemT in rebin
Switches to using template tag dispatching to determine ElemT in rebin.
This is the most readable, since we already have is_valueAndVariance_v
available to us.
I've added a comment too, in case anyone else feels
tempted to try and use constexpr if where they likely run into the
problem where we need the type to select the correct operator to get the
type.