Refactor to more canonical(?) use of transform
Note repeated work in bin_index_linspace kernel. We would like to pull
this out of the inner loop (or rather let the compiler do so). A
future/planned optimization in transform may modify the code in a way
such that the compiler will actually be able to do so.