make it easier to fetch data
For neutron tutorial and scipp-neutron and group-by scipp-neutron
examples, data comes from Mantid SystemTestData, cmake ExternalData.
I think the only easy way to get this without embarking on the Manid
cmake setup and build process is to download and rename as is currently
being done.

For the instrument view, we can get all the data we need from
TrainingCourseData zip. For that reason I have changed the file to not
be one that has to come through the ExternalData links.