Workaround around lack of cmake config from PyPI tbb-devel
This "patches" the cmake/TBB folder from the conda-forge tbb-devel
package into the pip install location of tbb-devel. Then we configure
TBB_DIR accordingly. This gets us through the cmake config and build.

To make a wheel, we then need to ensure that these libraries are found,
otherwise the "repair wheel" step of cibuildwheel fails. We use the new
(in cmake 3.21) IMPORTED_RUNTIME_ARTIFACTS option of `install`. This
means that our wheels will not contain the tbb shared libraries.