Fix performance bug with datetime64 operations
Many operations create "empties" as their output, which are then
overwritten with the results of the operations. It is generally assumed
that creating an empty is fast, since no initialization (not even
zeroing) is done.

It was observed that even simple operations such as `copy` were
significantly slower for datetime64. It turned out that a
member-initialization of the duration field to 0 was the culprit.
time_point was thus not a POD type, and C++ called the init code, even
when creating an "empty".

The result were 1000x slower "empty" calls, and 50x slower "copy" calls,
for example.