Use constraints instead of bound
Mypy does not recognize isinstance-based type checks when using bounds with a union in VariableLikeType. This is a problem in functions that check for the concrete type of their input to decide how to construct the output and where the output type is the same as the input type. It all works fine when using constraints instead of a bound.