MAINT: Do not delete nightlies from anaconda pypi
The current action has a bug as it will delete the nightly wheels if there wasn't a new commit to the main branch since the last nightly release.

Let's just turn it off for now. We can use the action upstream https://github.com/scientific-python/upload-nightly-action/pull/95 to remove the wheels once it's merged in.