Further reduce overhead in bin and group
Previously the mappers computed (and allocated variables) containing the
output bin sizes. These were then used to compute (and allocate
variables for) output bin indices. As the sizes themselves are not
needed, we can skip this and directly setup the indices. This saves
around 20% when grouping few rows into O(1M) bins.